+++ lib/pragma/macro/contract/persist.rb
@@ -5,20 +5,20 @@ require 'trailblazer/operation/persist'
 module Pragma
   module Macro
     module Contract
-      def self.Persist(**args)
+      def self.Persist(method: :save, name: "default")
         step = lambda do |input, options|
           Trailblazer::Operation::Pipetree::Step.new(
-            Trailblazer::Operation::Contract::Persist(**args).first
+            Trailblazer::Operation::Contract::Persist(method: method, name: name).first
           ).call(input, options).tap do |result|
             unless result
               options['result.response'] = Pragma::Operation::Response::UnprocessableEntity.new(
-                errors: options['model'].errors.messages
+                errors: options["contract.#{name}"].model.errors.messages
               ).decorate_with(Pragma::Decorator::Error)
             end
           end
         end
 
-        [step, name: 'persist.save']
+        [step, name: "contract.#{name}.#{method}"]
       end
     end
   end